requirements.txt
spacy==3.5.3
uvicorn==0.22.0
pymongo==4.4.0
pandas==1.3.5
fastapi==0.97.0
streamlit==1.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
openai==0.27.8
docker==6.1.3
faiss-cpu==1.7.4
python-dotenv==1.0.0